optimize roll option parsing
#2119
